Conflicts:
	Classes/Controller/RideController.php
	Resources/Private/Partials/Ride/FormFields.html
	Resources/Private/Templates/Ride/New.html

+++ Classes/Controller/RideController.php
@@ -2,7 +2,6 @@
 namespace Eike\Ride\Controller;
 
 use TYPO3\CMS\Extbase\Utility\DebuggerUtility;
-
 /***************************************************************
  *
  *  Copyright notice
@@ -41,6 +40,28 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      * @inject
      */
     protected $rideRepository = NULL;
+    /**
+     * addressRepository
+     *
+     * @var \Eike\Ride\Domain\Repository\AddressRepository
+     * @inject
+     */
+    protected $addressRepository = NULL;
+    
+    /**
+     * access
+     *
+     * @var \Eike\Ride\Service\Access
+     * @inject
+     */
+    protected $access = NULL;
+    
+    /**
+     *
+     * @var \Eike\Ride\Service\Address
+     * @inject
+     */
+    protected $addressService = NULL;
     
     public function initializeCreateAction() {
         if (isset($this->arguments['newRide'])) {
@@ -68,7 +89,10 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
     public function listAction()
     {
         $rides = $this->rideRepository->findAll();
+        
+        $this->view->assign('feUser', $this->access->getLoggedInFrontendUser());
         $this->view->assign('rides', $rides);
+        $this->view->assign('destination', $this->addressRepository->findByUid($this->settings['destination']));
     }
     
     /**
@@ -89,7 +113,10 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      */
     public function newAction()
     {
-        $this->view->assign('now', new \DateTime());
+    	if(!$this->access->getLoggedInFrontendUser()){
+    		throw new InsufficientUserPermissionsException('You are not logged in so you cannot create something here',1466258305);
+    	}
+    	$this->view->assign('driver', $this->access->getLoggedInFrontendUser());
     }
     
     /**
@@ -100,7 +127,8 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      */
     public function createAction(\Eike\Ride\Domain\Model\Ride $newRide)
     {
-        $this->addFlashMessage('The object was created. Please be aware that this action is publicly accessible unless you implement an access check. See http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+    	$this->addFlashMessage('The object was created.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
+    	$this->addressService->updateCoordinates($newRide->getStart());
         $this->rideRepository->add($newRide);
         $this->redirect('list');
     }
@@ -114,7 +142,11 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      */
     public function editAction(\Eike\Ride\Domain\Model\Ride $ride)
     {
+    	if(!$this->access->mayEditOrDelete($ride, $this->access->getLoggedInFrontendUser())){
+    		throw new InsufficientUserPermissionsException('You are not allowed to edit this ride',1466260533);
+    	}
         $this->view->assign('ride', $ride);
+        $this->view->assign('driver', $this->access->getLoggedInFrontendUser());
     }
     
     /**
@@ -125,7 +157,8 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      */
     public function updateAction(\Eike\Ride\Domain\Model\Ride $ride)
     {
-        $this->addFlashMessage('The object was updated. Please be aware that this action is publicly accessible unless you implement an access check. See http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+        $this->addFlashMessage('The object was updated.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
+        $this->addressService->updateCoordinates($ride->getStart());
         $this->rideRepository->update($ride);
         $this->redirect('list');
     }
@@ -138,7 +171,10 @@ class RideController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController
      */
     public function deleteAction(\Eike\Ride\Domain\Model\Ride $ride)
     {
-        $this->addFlashMessage('The object was deleted. Please be aware that this action is publicly accessible unless you implement an access check. See http://wiki.typo3.org/T3Doc/Extension_Builder/Using_the_Extension_Builder#1._Model_the_domain', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::ERROR);
+    	if(!$this->access->mayEditOrDelete($ride, $this->access->getLoggedInFrontendUser())){
+    		throw new InsufficientUserPermissionsException('You are not allowed to delete this ride',1466260675);
+    	}
+        $this->addFlashMessage('The object was deleted.', '', \TYPO3\CMS\Core\Messaging\AbstractMessage::OK);
         $this->rideRepository->remove($ride);
         $this->redirect('list');
     }
